uniclapp-backend/uniclapp/requirements.txt
Django==4.0
djangorestframework==3.12.4
django-extensions==3.1.3
pytz==2021.3
pillow==8.0.1
